chore: Overall improvements and bug fixes - Fixed float! to int! on cart query - Implemented product sorting - Implemented pagination getters (on facetGetters) - Aloow filtering on multiple attributes - Return attributes from cart query - Return product slug from cart query - Add product slug on all product queries - Implement slugGetter based on slug returned from API - Implement conditional chaining on product getters

+++ packages/api-client/src/api/cartStart/cartStart.ts
@@ -1,7 +1,7 @@
 import gql from 'graphql-tag';
 
 export default gql`
-  mutation cartStart($siteId: Float!) {
+  mutation cartStart($siteId: Int!) {
     cartStart(siteId: $siteId) {
       cartId
     }
